refactor: Sprint 48 PR 2 — split telegram.rs core (3a)
Extract 5 sub-modules from telegram/mod.rs (core transport + inbound):
- state.rs (171 LOC): TelegramState, lock_state, runtime helpers
- topic_registry.rs (120 LOC): topic CRUD, create/delete/lookup
- send.rs (153 LOC): send primitives, media handling
- inbound.rs (474 LOC): start_polling, handle_message
- error.rs (151 LOC): topic-deleted detection, cleanup, self-heal

mod.rs retains PR 3b content (creds, reply, bot_api, notify, adapter,
bootstrap) + tests — will be extracted in next PR.

Zero behavior change. All pub(crate) re-exported via mod.rs.
