+++ src/channel/telegram/mod.rs
@@ -3,6 +3,18 @@
 //! Inbound: Telegram message → inbox + PTY notification
 //! Outbound: reply(text) → Telegram send_message to topic
 
+pub(crate) mod error;
+pub(crate) mod inbound;
+pub(crate) mod send;
+pub(crate) mod state;
+pub(crate) mod topic_registry;
+
+pub(crate) use error::*;
+pub(crate) use inbound::*;
+pub(crate) use send::*;
+pub(crate) use state::*;
+pub(crate) use topic_registry::*;
+
 use crate::agent::AgentRegistry;
 use crate::fleet::ChannelConfig;
 use crate::inbox::{self, InboxMessage};
@@ -13,988 +25,6 @@ use std::sync::{Arc, OnceLock};
 use teloxide::net::Download;
 use teloxide::prelude::*;
 use teloxide::types::{MessageId, ThreadId};
-
-/// Lock TelegramState, recovering from poison.
-/// With parking_lot::Mutex, lock never fails (no poisoning).
-pub(crate) fn lock_state(
-    tg: &Arc<Mutex<TelegramState>>,
-) -> parking_lot::MutexGuard<'_, TelegramState> {
-    tg.lock()
-}
-
-// ---------------------------------------------------------------------------
-// Topic registry — persists topic_id → instance_name in $AGEND_HOME/topics.json
-// so we can detect orphaned topics on daemon restart.
-// ---------------------------------------------------------------------------
-
-/// Reserved pseudo-instance name used in `topics.json` to pin the
-/// `fleet_binding` topic across daemon restarts. Not a real instance —
-/// chosen so it can never collide with a user-configured name
-/// (`fleet.yaml` keys are slugs; underscores-bracketing is reserved).
-/// See [`init_from_config`] orphan-cleanup filter and fleet-binding
-/// resolution.
-const FLEET_BINDING_SENTINEL: &str = "__fleet__";
-
-fn topic_registry_path(home: &Path) -> PathBuf {
-    home.join("topics.json")
-}
-
-fn load_topic_registry(home: &Path) -> HashMap<i32, String> {
-    let path = topic_registry_path(home);
-    std::fs::read_to_string(&path)
-        .ok()
-        .and_then(|s| serde_json::from_str::<HashMap<String, String>>(&s).ok())
-        .map(|m| {
-            m.into_iter()
-                .filter_map(|(k, v)| k.parse::<i32>().ok().map(|id| (id, v)))
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn save_topic_registry(home: &Path, registry: &HashMap<i32, String>) {
-    let map: HashMap<String, &String> = registry.iter().map(|(k, v)| (k.to_string(), v)).collect();
-    if let Ok(json) = serde_json::to_string_pretty(&map) {
-        // H2: atomic write to prevent partial-file on crash
-        let _ = crate::store::atomic_write(&topic_registry_path(home), json.as_bytes());
-    }
-}
-
-fn register_topic(home: &Path, topic_id: i32, instance_name: &str) {
-    // Write-side unification (Fix B): all three sources updated atomically.
-    // 1. topics.json (disk registry)
-    let mut reg = load_topic_registry(home);
-    reg.insert(topic_id, instance_name.to_string());
-    save_topic_registry(home, &reg);
-    // 2. fleet.yaml topic_id field
-    let _ = crate::fleet::update_instance_field(
-        home,
-        instance_name,
-        "topic_id",
-        serde_yaml_ng::Value::Number(serde_yaml_ng::Number::from(topic_id)),
-    );
-    // 3. in-memory state is updated by the caller (Channel trait methods
-    //    that hold &self.state). Free-function callers without state access
-    //    rely on resolve_topic's topics.json fallback as defense-in-depth.
-}
-
-fn unregister_topic(home: &Path, topic_id: i32) {
-    let mut reg = load_topic_registry(home);
-    reg.remove(&topic_id);
-    save_topic_registry(home, &reg);
-}
-
-/// Shared tokio runtime for all Telegram sync→async calls.
-fn telegram_runtime() -> &'static tokio::runtime::Runtime {
-    static RT: OnceLock<tokio::runtime::Runtime> = OnceLock::new();
-    RT.get_or_init(|| {
-        tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .expect("telegram tokio runtime")
-    })
-}
-
-/// Run a future on the Telegram runtime. If already inside an async context
-/// (e.g. Telegram polling → emit path), spawns a fire-and-forget task on the
-/// current runtime to avoid `block_on`-inside-runtime panic. Returns `Ok(())`
-/// for the spawned path since the result is not awaited.
-///
-/// H3: the spawn path returns Ok(()) immediately — errors are logged but not
-/// propagated. This is intentional: the caller is in a sync context and cannot
-/// await the spawned task. Callers must not assume Ok(()) means the operation
-/// succeeded — it means the task was submitted. Check tracing logs for failures.
-fn spawn_or_block_on<F>(fut: F) -> anyhow::Result<()>
-where
-    F: std::future::Future<Output = anyhow::Result<()>> + Send + 'static,
-{
-    if let Ok(handle) = tokio::runtime::Handle::try_current() {
-        handle.spawn(async move {
-            if let Err(e) = fut.await {
-                tracing::warn!(%e, "telegram spawn task failed");
-            }
-        });
-        Ok(())
-    } else {
-        telegram_runtime().block_on(fut)
-    }
-}
-
-pub struct TelegramState {
-    /// `None` only inside the contract-test harness — production `new`
-    /// always populates it via `Bot::new`. Transport methods unwrap with
-    /// `.expect("telegram bot not initialized")`; contract tests never
-    /// reach those paths (see `src/channel/contract.rs` scope comment).
-    pub bot: Option<Bot>,
-    #[allow(dead_code)]
-    pub group_id: ChatId,
-    pub topic_to_instance: HashMap<i32, String>,
-    #[allow(dead_code)]
-    pub instance_to_topic: HashMap<String, i32>,
-    pub home: PathBuf,
-    /// Submit key per instance (for PTY notification injection).
-    pub submit_keys: HashMap<String, String>,
-    /// Allowlist of Telegram user IDs permitted to command the fleet.
-    /// See [`crate::fleet::ChannelConfig::Telegram::user_allowlist`] for
-    /// semantics of `None` vs `Some(empty)` vs `Some([...])`.
-    pub user_allowlist: Option<Vec<i64>>,
-    /// Wired in post-bootstrap by [`attach_registry`]; lets inbound message
-    /// routing read `agent_state` directly instead of via the `LIST` RPC.
-    pub registry: Option<AgentRegistry>,
-    /// Resolved `fleet_binding` target for cross-instance fleet activity
-    /// rendering (Stage B-UX, `docs/archived/DESIGN-stage-b-ux.md` §3/§5). `None`
-    /// means no mirror is configured — [`TelegramChannel::apply_fleet_action`]
-    /// returns early. Resolution happens in [`init_from_config`] from the
-    /// config's `fleet_binding` block plus the on-disk topic registry
-    /// sentinel `"__fleet__"`.
-    pub fleet_binding_topic_id: Option<i32>,
-}
-
-impl TelegramState {
-    pub fn new(
-        token: &str,
-        group_id: i64,
-        topic_map: HashMap<String, i32>,
-        home: PathBuf,
-        submit_keys: HashMap<String, String>,
-        user_allowlist: Option<Vec<i64>>,
-    ) -> Self {
-        let topic_to_instance: HashMap<i32, String> = topic_map
-            .iter()
-            .map(|(name, &tid)| (tid, name.clone()))
-            .collect();
-        Self {
-            bot: Some(Bot::new(token)),
-            group_id: ChatId(group_id),
-            topic_to_instance,
-            instance_to_topic: topic_map,
-            home,
-            submit_keys,
-            user_allowlist,
-            registry: None,
-            fleet_binding_topic_id: None,
-        }
-    }
-
-    /// Build a `TelegramState` without constructing a `teloxide::Bot` —
-    /// used by the `src/channel/contract.rs` harness, which only exercises
-    /// registry-side methods (`kind`, `has_binding`, `take_binding`,
-    /// `record_binding`, `attach_registry`). `Bot::new` eagerly initializes
-    /// reqwest + `system-configuration` proxy state and panics on some
-    /// macOS setups, so the harness must not go through it. If a test
-    /// triggers a transport path (`send_to_topic`, `send_reply`, polling),
-    /// the `.expect("telegram bot not initialized")` unwrap will fire.
-    #[cfg(test)]
-    pub(crate) fn new_for_contract_test(
-        group_id: i64,
-        topic_map: HashMap<String, i32>,
-        home: PathBuf,
-        submit_keys: HashMap<String, String>,
-        user_allowlist: Option<Vec<i64>>,
-    ) -> Self {
-        let topic_to_instance: HashMap<i32, String> = topic_map
-            .iter()
-            .map(|(name, &tid)| (tid, name.clone()))
-            .collect();
-        Self {
-            bot: None,
-            group_id: ChatId(group_id),
-            topic_to_instance,
-            instance_to_topic: topic_map,
-            home,
-            submit_keys,
-            user_allowlist,
-            registry: None,
-            fleet_binding_topic_id: None,
-        }
-    }
-
-    /// Return true if a sender is permitted by the allowlist.
-    ///
-    /// **Sprint 21 Phase 2 fail-closed swap (PR #216 + #217 cascade
-    /// auth)**: previously `None` allowlist returned `true` (legacy
-    /// accept-all); now it returns `false`. The implementation
-    /// delegates to `crate::channel::auth::is_authorized_recipient`,
-    /// which is the single source of truth shared with Phase 1's
-    /// outbound notify gate. Operators must configure
-    /// `user_allowlist: [user_id, ...]` in `fleet.yaml` to enable
-    /// inbound (and outbound) traffic — see `docs/USAGE.md` "Channel:
-    /// Telegram" section for the migration steps.
-    pub fn is_user_allowed(&self, user_id: i64) -> bool {
-        crate::channel::auth::is_authorized_recipient(&self.user_allowlist, user_id)
-    }
-
-    /// Send a message to an instance's Telegram topic.
-    #[allow(dead_code)]
-    pub async fn send_to_topic(&self, instance_name: &str, text: &str) -> anyhow::Result<()> {
-        let topic_id = self
-            .instance_to_topic
-            .get(instance_name)
-            .ok_or_else(|| anyhow::anyhow!("No topic for '{instance_name}'"))?;
-        let bot = self
-            .bot
-            .as_ref()
-            .expect("telegram bot not initialized (contract-test construction?)");
-        send_with_topic(bot, self.group_id, Some(*topic_id), text, None).await
-    }
-}
-
-/// Send a message, optionally to a topic, optionally as a reply, returning
-/// the platform-assigned `message_id` (i32 per Telegram Bot API).
-///
-/// Phase 3 (Sprint 21) extracts msg_id capture so `Channel::send` can
-/// return a `MsgRef` with a real id instead of `"0"` placeholder. Other
-/// callers that historically returned `Result<()>` continue to use
-/// [`send_with_topic`] (thin wrapper) so the behaviour change is
-/// localised to the trait-method dispatch path.
-async fn send_with_topic_capturing_id(
-    bot: &Bot,
-    chat_id: ChatId,
-    topic_id: Option<i32>,
-    text: &str,
-    reply_to_msg_id: Option<i32>,
-) -> anyhow::Result<i32> {
-    use teloxide::payloads::SendMessageSetters;
-    use teloxide::prelude::Requester;
-    use teloxide::types::ReplyParameters;
-    let mut req = bot.send_message(chat_id, text);
-    if let Some(tid) = topic_id {
-        if tid != 1 {
-            req = req.message_thread_id(ThreadId(MessageId(tid)));
-        }
-    }
-    if let Some(mid) = reply_to_msg_id {
-        req = req.reply_parameters(ReplyParameters::new(MessageId(mid)));
-    }
-    let sent = req.send().await?;
-    Ok(sent.id.0)
-}
-
-/// Backwards-compatible thin wrapper around
-/// [`send_with_topic_capturing_id`] for callers that don't need the
-/// returned `message_id`. Existing five call sites
-/// (`send_to_topic` / `notify_telegram_inner` / `apply_fleet_action` /
-/// the text-after-attachment follow-up in `Channel::send` /
-/// `try_telegram_*` etc.) keep their `Result<()>` shape.
-async fn send_with_topic(
-    bot: &Bot,
-    chat_id: ChatId,
-    topic_id: Option<i32>,
-    text: &str,
-    reply_to_msg_id: Option<i32>,
-) -> anyhow::Result<()> {
-    send_with_topic_capturing_id(bot, chat_id, topic_id, text, reply_to_msg_id)
-        .await
-        .map(|_| ())
-}
-
-/// Telegram Bot API caption limit (characters).
-const CAPTION_MAX_CHARS: usize = 1024;
-
-/// Resolve the caption for a media send.
-fn resolve_caption(text: &str, att: &crate::channel::Attachment) -> Option<String> {
-    use crate::channel::AttachmentKind;
-    if att.kind == AttachmentKind::Sticker {
-        return None;
-    }
-    if let Some(ref cap) = att.caption {
-        return Some(cap.chars().take(CAPTION_MAX_CHARS).collect());
-    }
-    if !text.is_empty() && text.chars().count() <= CAPTION_MAX_CHARS {
-        return Some(text.to_string());
-    }
-    None
-}
-
-/// Returns true when text should be sent as a separate follow-up message.
-fn needs_separate_text(text: &str, att: &crate::channel::Attachment) -> bool {
-    use crate::channel::AttachmentKind;
-    if text.is_empty() {
-        return false;
-    }
-    if att.caption.is_some() {
-        return true;
-    }
-    att.kind == AttachmentKind::Sticker || text.chars().count() > CAPTION_MAX_CHARS
-}
-
-/// Send a media attachment via the appropriate Telegram Bot API method.
-async fn send_media(
-    bot: &Bot,
-    chat_id: ChatId,
-    topic_id: Option<i32>,
-    att: &crate::channel::Attachment,
-    caption: Option<&str>,
-) -> anyhow::Result<i32> {
-    use crate::channel::AttachmentKind;
-    use teloxide::prelude::Requester;
-    use teloxide::types::InputFile;
-
-    let input = InputFile::file(&att.path);
-    let input = if let Some(ref name) = att.original_filename {
-        input.file_name(name.clone())
-    } else {
-        input
-    };
-    let thread = topic_id.filter(|&t| t != 1).map(|t| ThreadId(MessageId(t)));
-
-    let msg_id = match att.kind {
-        AttachmentKind::Photo => {
-            let mut req = bot.send_photo(chat_id, input);
-            if let Some(cap) = caption {
-                req = req.caption(cap);
-            }
-            if let Some(tid) = thread {
-                req = req.message_thread_id(tid);
-            }
-            req.await?.id.0
-        }
-        AttachmentKind::Voice => {
-            let mut req = bot.send_voice(chat_id, input);
-            if let Some(cap) = caption {
-                req = req.caption(cap);
-            }
-            if let Some(tid) = thread {
-                req = req.message_thread_id(tid);
-            }
-            req.await?.id.0
-        }
-        AttachmentKind::Document => {
-            let mut req = bot.send_document(chat_id, input);
-            if let Some(cap) = caption {
-                req = req.caption(cap);
-            }
-            if let Some(tid) = thread {
-                req = req.message_thread_id(tid);
-            }
-            req.await?.id.0
-        }
-        AttachmentKind::Video => {
-            let mut req = bot.send_video(chat_id, input);
-            if let Some(cap) = caption {
-                req = req.caption(cap);
-            }
-            if let Some(tid) = thread {
-                req = req.message_thread_id(tid);
-            }
-            req.await?.id.0
-        }
-        AttachmentKind::Sticker => {
-            let mut req = bot.send_sticker(chat_id, input);
-            if let Some(tid) = thread {
-                req = req.message_thread_id(tid);
-            }
-            req.await?.id.0
-        }
-    };
-    Ok(msg_id)
-}
-
-/// Start Telegram polling in a dedicated thread with its own tokio runtime.
-pub fn start_polling(state: Arc<Mutex<TelegramState>>) {
-    // fire-and-forget: telegram polling thread runs the teloxide dispatcher
-    // for the daemon's lifetime. Stops when the bot's update stream errors
-    // (network drop / shutdown). No JoinHandle / shutdown signal needed —
-    // process exit reaps the thread.
-    if let Err(e) = std::thread::Builder::new()
-        .name("telegram".into())
-        .spawn(move || {
-            let _census = crate::thread_census::register("telegram_poll");
-            let Ok(rt) = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()
-            else {
-                tracing::error!("failed to build tokio runtime");
-                return;
-            };
-            rt.block_on(async {
-                let bot = lock_state(&state)
-                    .bot
-                    .clone()
-                    .expect("telegram bot not initialized (polling thread)");
-                let state2 = Arc::clone(&state);
-                let handler = Update::filter_message().endpoint(move |_bot: Bot, msg: Message| {
-                    let state = Arc::clone(&state2);
-                    async move {
-                        handle_message(&state, &msg).await;
-                        respond(())
-                    }
-                });
-                tracing::info!("polling started");
-                Dispatcher::builder(bot, handler).build().dispatch().await;
-            });
-        })
-    {
-        tracing::error!(error = %e, "failed to spawn polling thread");
-    }
-}
-
-/// Resolve a topic_id to an instance name.
-/// First checks the in-memory map, then reloads from fleet.yaml for
-/// runtime-created topics (via create_instance).
-fn resolve_topic(state: &mut TelegramState, topic_id: Option<i32>) -> String {
-    if let Some(tid) = topic_id {
-        if let Some(name) = state.topic_to_instance.get(&tid).cloned() {
-            return name;
-        }
-        // Unknown topic_id — reload from fleet.yaml
-        if let Ok(config) = crate::fleet::FleetConfig::load(&state.home.join("fleet.yaml")) {
-            for (inst_name, inst) in &config.instances {
-                if inst.topic_id == Some(tid) {
-                    state.topic_to_instance.insert(tid, inst_name.clone());
-                    state.instance_to_topic.insert(inst_name.clone(), tid);
-                    return inst_name.clone();
-                }
-            }
-        }
-        // Fix B: check topics.json as third source (runtime-created topics
-        // may exist there but not in fleet.yaml or in-memory state).
-        let reg = load_topic_registry(&state.home);
-        if let Some(inst_name) = reg.get(&tid) {
-            state.topic_to_instance.insert(tid, inst_name.clone());
-            state.instance_to_topic.insert(inst_name.clone(), tid);
-            return inst_name.clone();
-        }
-        // Fix C: warn before fallback so operator sees misroute.
-        tracing::warn!(
-            thread_id = tid,
-            "resolve_topic: topic_id not found in memory, fleet.yaml, or topics.json — falling back to \"general\""
-        );
-    }
-    "general".to_string()
-}
-
-async fn handle_message(state: &Arc<Mutex<TelegramState>>, msg: &Message) {
-    // Detect topic closure/deletion — auto-delete the corresponding instance
-    if msg.forum_topic_closed().is_some() {
-        let thread_id = msg.thread_id.map(|ThreadId(MessageId(id))| id);
-        if let Some(tid) = thread_id {
-            let (instance_name, home) = {
-                let s = lock_state(state);
-                (s.topic_to_instance.get(&tid).cloned(), s.home.clone())
-            };
-            match instance_name {
-                Some(name) => {
-                    tracing::info!(topic_id = tid, instance = %name, "topic closed, deleting instance");
-                    cleanup_deleted_topic(&home, &name, tid, Some(state));
-                }
-                None => tracing::warn!(topic_id = tid, "topic closed (no matching instance)"),
-            }
-            return;
-        }
-        tracing::warn!("topic closed (no thread_id)");
-        return;
-    }
-
-    let text = match msg.text() {
-        Some(t) => t.to_string(),
-        None => msg.caption().unwrap_or("").to_string(),
-    };
-
-    // Status keyword detection — surface summary to the resolved instance's inbox.
-    // The agent (typically general) sees the summary and can relay to operator.
-    if crate::status_summary::is_status_keyword(&text) {
-        let thread_id = msg.thread_id.map(|ThreadId(MessageId(id))| id);
-        let instance_name = {
-            let mut s = lock_state(state);
-            resolve_topic(&mut s, thread_id)
-        };
-        let home = lock_state(state).home.clone();
-        let summary = crate::status_summary::build_summary(&home);
-        tracing::info!(to = %instance_name, "status keyword detected, injecting summary");
-        let _ = crate::inbox::enqueue(
-            &home,
-            &instance_name,
-            crate::inbox::InboxMessage {
-                schema_version: 0,
-                id: None,
-                read_at: None,
-                thread_id: None,
-                parent_id: None,
-                task_id: None,
-                force_meta: None,
-                correlation_id: None,
-                reviewed_head: None,
-                from: "system:status".to_string(),
-                text: summary,
-                kind: Some("status-summary".to_string()),
-                timestamp: chrono::Utc::now().to_rfc3339(),
-                channel: Some(crate::channel::ChannelKind::Telegram),
-                delivery_mode: None,
-                attachments: vec![],
-                in_reply_to_msg_id: None,
-                in_reply_to_excerpt: None,
-                superseded_by: None,
-                from_id: None,
-            },
-        );
-        // Also notify agent PTY so it picks up the summary
-        let username = msg
-            .from
-            .as_ref()
-            .and_then(|u| u.username.as_deref())
-            .unwrap_or("unknown");
-        crate::inbox::notify_agent(
-            &home,
-            &instance_name,
-            &crate::inbox::NotifySource::Channel(username, crate::channel::ChannelKind::Telegram),
-            "[status-summary] check inbox for status overview",
-        );
-        return;
-    }
-
-    // Task entry via telegram: "加 task: <title>" creates a task assigned to dev-lead.
-    if let Some(title) = crate::status_summary::parse_task_entry(&text) {
-        let home = lock_state(state).home.clone();
-        let result = crate::tasks::handle(
-            &home,
-            "operator",
-            &serde_json::json!({
-                "action": "create",
-                "title": title,
-                "assignee": "dev-lead",
-                "priority": "normal",
-            }),
-        );
-        let task_id = result["id"].as_str().unwrap_or("?");
-        tracing::info!(title, task_id, "task created via telegram keyword");
-        if let Some(ch) = crate::channel::active_channel() {
-            let _ = ch.notify(
-                "operator",
-                crate::channel::NotifySeverity::Info,
-                &format!("✅ Task created: {title} [{task_id}]"),
-                false,
-            );
-        }
-        return;
-    }
-
-    // Extract inbound attachment metadata (photo/voice/document/video/sticker).
-    // Download happens later after topic resolution provides instance_name.
-    struct InboundFile<'a> {
-        file_id: &'a str,
-        kind: crate::channel::event::AttachmentKind,
-        mime: Option<String>,
-        size: Option<u64>,
-        filename: Option<String>,
-    }
-    let inbound_file: Option<InboundFile<'_>> = {
-        use crate::channel::event::AttachmentKind;
-        if let Some(sizes) = msg.photo() {
-            sizes.last().map(|p| InboundFile {
-                file_id: p.file.id.as_str(),
-                kind: AttachmentKind::Photo,
-                mime: None,
-                size: Some(p.file.size as u64),
-                filename: None,
-            })
-        } else if let Some(doc) = msg.document() {
-            Some(InboundFile {
-                file_id: doc.file.id.as_str(),
-                kind: AttachmentKind::Document,
-                mime: doc.mime_type.as_ref().map(|m| m.to_string()),
-                size: Some(doc.file.size as u64),
-                filename: doc.file_name.clone(),
-            })
-        } else if let Some(voice) = msg.voice() {
-            Some(InboundFile {
-                file_id: voice.file.id.as_str(),
-                kind: AttachmentKind::Voice,
-                mime: voice.mime_type.as_ref().map(|m| m.to_string()),
-                size: Some(voice.file.size as u64),
-                filename: None,
-            })
-        } else if let Some(video) = msg.video() {
-            Some(InboundFile {
-                file_id: video.file.id.as_str(),
-                kind: AttachmentKind::Video,
-                mime: video.mime_type.as_ref().map(|m| m.to_string()),
-                size: Some(video.file.size as u64),
-                filename: video.file_name.clone(),
-            })
-        } else {
-            msg.sticker().map(|sticker| InboundFile {
-                file_id: sticker.file.id.as_str(),
-                kind: AttachmentKind::Sticker,
-                mime: None,
-                size: Some(sticker.file.size as u64),
-                filename: None,
-            })
-        }
-    };
-
-    // If no text and no attachment, nothing to process.
-    if text.is_empty() && inbound_file.is_none() {
-        return;
-    }
-
-    let sender_id: Option<i64> = msg.from.as_ref().map(|u| u.id.0 as i64);
-    let username = msg
-        .from
-        .as_ref()
-        .and_then(|u| u.username.as_deref())
-        .unwrap_or("unknown");
-
-    // Authz: drop messages from senders not on the allowlist (Sprint 21
-    // Phase 2 fail-closed cascade auth — combined with PR #216 outbound
-    // gate this closes the cascade attack chain inbound side).
-    //   - `Some([u1, u2])` → restricts to listed IDs (always was)
-    //   - `Some([])` → rejects all
-    //   - `None` → rejects all (Phase 2 inversion of legacy `accept-all`)
-    //   - sender_id `None` (anonymous) → fail-closed: rejected
-    {
-        let s = lock_state(state);
-        let allowed = match sender_id {
-            Some(id) => s.is_user_allowed(id),
-            None => false,
-        };
-        if !allowed {
-            tracing::warn!(
-                from = username,
-                user_id = ?sender_id,
-                "telegram message rejected by user_allowlist"
-            );
-            return;
-        }
-    }
-
-    let thread_id = msg.thread_id.map(|ThreadId(MessageId(id))| id);
-
-    let (instance_name, home, _submit_key, registry) = {
-        let mut s = lock_state(state);
-        let name = resolve_topic(&mut s, thread_id);
-        let sk = s
-            .submit_keys
-            .get(&name)
-            .cloned()
-            .unwrap_or_else(|| "\r".to_string());
-        (name, s.home.clone(), sk, s.registry.clone())
-    };
-
-    tracing::info!(from = username, to = %instance_name, %text, "inbound message");
-
-    // Route based on agent state: when blocked on an interactive prompt
-    // (AwaitingOperator startup stall, or a pattern-matched InteractivePrompt
-    // like codex's update menu), the operator's reply must reach the PTY as
-    // raw keystrokes — any inbox prefix ("[telegram:@user] …") would confuse
-    // the CLI's prompt parser. In every other state, preserve the existing
-    // inbox semantics so agent-authored message handling keeps working.
-    if agent_wants_raw_keystrokes(registry.as_ref(), &instance_name) {
-        let payload = format!("{text}\n");
-        match crate::api::call(
-            &home,
-            &serde_json::json!({
-                "method": crate::api::method::INJECT,
-                "params": {"name": instance_name, "data": payload, "raw": true}
-            }),
-        ) {
-            Ok(_) => tracing::info!(
-                to = %instance_name,
-                bytes = payload.len(),
-                "routed raw keystrokes (interactive prompt)"
-            ),
-            Err(e) => tracing::warn!(
-                to = %instance_name,
-                error = %e,
-                "raw injection failed"
-            ),
-        }
-        return;
-    }
-
-    // Persist the inbound message ID so the UX layer can resolve the
-    // origin message when emitting AgentPickedUp (channel-agnostic).
-    // Append to pending_pickup_ids array so multi-message bursts each
-    // get a ✅ confirmation on inbox drain (F2 fix).
-    {
-        let meta_path = home.join("metadata").join(format!("{instance_name}.json"));
-        let mut meta: serde_json::Value = std::fs::read_to_string(&meta_path)
-            .ok()
-            .and_then(|c| serde_json::from_str(&c).ok())
-            .unwrap_or(serde_json::json!({}));
-        let entry = serde_json::json!({
-            "kind": "telegram",
-            "msg_id": msg.id.0.to_string(),
-        });
-        match meta.get_mut("pending_pickup_ids") {
-            Some(arr) if arr.is_array() => {
-                let vec = arr.as_array_mut().expect("checked");
-                vec.push(entry);
-                // M2: cap to prevent unbounded growth (keep newest 100)
-                if vec.len() > 100 {
-                    *vec = vec.split_off(vec.len() - 100);
-                }
-            }
-            _ => {
-                meta["pending_pickup_ids"] = serde_json::json!([entry]);
-            }
-        }
-        let meta_dir = home.join("metadata");
-        std::fs::create_dir_all(&meta_dir).ok();
-        let tmp_path = meta_path.with_extension("json.tmp");
-        if let Ok(content) = serde_json::to_string_pretty(&meta) {
-            if std::fs::write(&tmp_path, &content).is_ok() {
-                let _ = std::fs::rename(&tmp_path, &meta_path);
-            }
-        }
-    }
-    // Also store as last_message_id for the `react` MCP tool's fallback.
-    crate::agent_ops::save_metadata(
-        &home,
-        &instance_name,
-        "last_message_id",
-        serde_json::json!(msg.id.0),
-    );
-
-    // Download inbound attachment if present (async — avoids nested runtime panic).
-    let attachments: Vec<crate::channel::event::Attachment> = if let Some(f) = inbound_file {
-        let bot = lock_state(state).bot.clone();
-        let result = match bot {
-            Some(bot) => download_file_async(&bot, &home, &instance_name, f.file_id).await,
-            None => Err(anyhow::anyhow!("telegram bot not initialized")),
-        };
-        match result {
-            Ok(local_path) => vec![crate::channel::event::Attachment {
-                kind: f.kind,
-                path: std::path::PathBuf::from(&local_path),
-                mime: f.mime,
-                caption: msg.caption().map(String::from),
-                size_bytes: f.size,
-                original_filename: f.filename,
-            }],
-            Err(e) => {
-                tracing::warn!(file_id = f.file_id, error = %e, "inbound attachment download failed");
-                vec![]
-            }
-        }
-    } else {
-        vec![]
-    };
-
-    // Enqueue in inbox
-    let msg_obj = InboxMessage {
-        schema_version: 0,
-        id: None,
-        read_at: None,
-        thread_id: None,
-        parent_id: None,
-        task_id: None,
-        force_meta: None,
-        correlation_id: None,
-        reviewed_head: None,
-        from: format!("user:{username}"),
-        text: text.to_string(),
-        kind: None, // was "telegram" — channel source now in typed `channel` field
-        timestamp: chrono::Utc::now().to_rfc3339(),
-        channel: Some(crate::channel::ChannelKind::Telegram),
-        delivery_mode: None,
-        attachments,
-        in_reply_to_msg_id: msg.reply_to_message().map(|r| r.id.0.to_string()),
-        in_reply_to_excerpt: msg.reply_to_message().and_then(|r| {
-            let text = r.text().or_else(|| r.caption()).unwrap_or("");
-            let author = r
-                .from
-                .as_ref()
-                .and_then(|u| u.username.as_deref())
-                .unwrap_or("unknown");
-            inbox::build_excerpt(text, author)
-        }),
-        superseded_by: None,
-        from_id: None,
-    };
-    let _ = inbox::enqueue(&home, &instance_name, msg_obj);
-
-    // Notify agent PTY
-    inbox::notify_agent(
-        &home,
-        &instance_name,
-        &inbox::NotifySource::Channel(username, crate::channel::ChannelKind::Telegram),
-        &text,
-    );
-
-    // Emit UxEvent::UserMsgReceived so the channel adapter can react 👀.
-    {
-        use crate::channel::binding::BindingRef;
-        use crate::channel::event::MsgRef;
-        use crate::channel::ux_event::UxEvent;
-        let origin_msg = MsgRef {
-            binding: BindingRef::new("telegram", Some(instance_name.clone()), ()),
-            id: msg.id.0.to_string(),
-        };
-        crate::channel::sink_registry::registry().emit(&UxEvent::UserMsgReceived {
-            origin_msg,
-            agent: instance_name,
-        });
-    }
-}
-
-/// Classify a send error as "the bound topic was deleted out from under us".
-///
-/// Bot API 6.3+ exposes no typed variant or deletion service message for
-/// "thread gone"; it surfaces as `ApiError::Unknown("Bad Request: message
-/// thread not found")`. Substring match on the flattened chain tolerates both
-/// `anyhow::context` wrapping and future teloxide wrapping changes.
-const TOPIC_DELETED_MARKER: &str = "message thread not found";
-
-pub(crate) fn is_topic_deleted_error(err: &anyhow::Error) -> bool {
-    let s = format!("{err:#}").to_lowercase();
-    s.contains(TOPIC_DELETED_MARKER)
-}
-
-/// Cleanup path when a topic is known-deleted.
-///
-/// `state` is `None` for callers that lack access to `TelegramState` (daemon
-/// `notify_telegram`, `try_telegram_reply`). Those paths leave the in-memory
-/// maps stale until the next state-aware send or process restart — acceptable
-/// because the MCP `delete_instance` handler's disk mutation (fleet.yaml
-/// removal + `cleanup_working_dir`) is the source of truth.
-pub(crate) fn cleanup_deleted_topic(
-    home: &Path,
-    instance_name: &str,
-    tid: i32,
-    state: Option<&Arc<Mutex<TelegramState>>>,
-) {
-    if let Some(state) = state {
-        let mut s = lock_state(state);
-        s.topic_to_instance.remove(&tid);
-        s.instance_to_topic.remove(instance_name);
-        s.submit_keys.remove(instance_name);
-    }
-    let _ = crate::api::call(
-        home,
-        &serde_json::json!({"method": crate::api::method::DELETE, "params": {"name": instance_name}}),
-    );
-    if let Err(e) = crate::fleet::remove_instance_from_yaml(home, instance_name) {
-        tracing::warn!(instance = %instance_name, error = %e, "failed to remove from fleet.yaml");
-    }
-    // Strip the stale topic-id entry from the local registry. Safe to call
-    // even if the topic was already unregistered (HashMap::remove is a no-op
-    // on missing keys).
-    unregister_topic(home, tid);
-}
-
-/// Cleanup path for a deleted `fleet_binding` topic. Unlike
-/// [`cleanup_deleted_topic`] — which is instance-oriented and tears down
-/// fleet.yaml / api::DELETE / per-instance state maps — the fleet binding
-/// has no instance to drop, just a sentinel registry row and the
-/// `fleet_binding_topic_id` field.
-///
-/// Only clears `fleet_binding_topic_id` when it still points at `tid`
-/// (defensive — avoids clobbering a fresh binding if a stale error
-/// somehow arrives after re-resolution).
-pub(crate) fn cleanup_fleet_binding(home: &Path, state: &Arc<Mutex<TelegramState>>, tid: i32) {
-    unregister_topic(home, tid);
-    let mut s = lock_state(state);
-    if s.fleet_binding_topic_id == Some(tid) {
-        s.fleet_binding_topic_id = None;
-    }
-}
-
-/// Classify a fleet-binding send error and run [`cleanup_fleet_binding`]
-/// if it matches a topic-deleted error. Returns `true` when the error
-/// was handled (topic gone); the renderer uses that to silence the outer
-/// "send failed" warn since self-heal is the expected outcome, not a
-/// surprise.
-///
-/// Reviewer context (at-dev-4 on PR #56): without this, deleting the
-/// fleet topic once would (a) silently drop every subsequent fleet
-/// emission, and (b) persist the stale `__fleet__` row in `topics.json`
-/// so the next daemon restart happily reused the dead thread id.
-/// Regression pin: [`tests::fleet_binding_self_heals_when_topic_deleted`].
-pub(crate) fn handle_fleet_send_failure(
-    err: &anyhow::Error,
-    home: &Path,
-    state: &Arc<Mutex<TelegramState>>,
-    tid: i32,
-) -> bool {
-    if !is_topic_deleted_error(err) {
-        return false;
-    }
-    tracing::info!(
-        topic_id = tid,
-        "fleet send hit topic_deleted — clearing binding + unregistering sentinel"
-    );
-    cleanup_fleet_binding(home, state, tid);
-    true
-}
-
-/// Classify a send error and run topic-delete cleanup if it matches.
-/// Returns `true` when the error was handled (topic gone); callers may then
-/// silence the outer "send failed" log since cleanup is the expected outcome.
-pub(crate) fn handle_send_failure(
-    err: &anyhow::Error,
-    home: &Path,
-    instance_name: &str,
-    topic_id: Option<i32>,
-    state: Option<&Arc<Mutex<TelegramState>>>,
-) -> bool {
-    if !is_topic_deleted_error(err) {
-        return false;
-    }
-    let Some(tid) = topic_id else {
-        return false;
-    };
-    tracing::info!(
-        instance = %instance_name,
-        topic_id = tid,
-        "send hit topic_deleted — cleaning up"
-    );
-    cleanup_deleted_topic(home, instance_name, tid, state);
-    true
-}
-
-/// Lightweight self-heal for a stale topic: strip the dead topic_id from
-/// the on-disk registry and fleet.yaml, create a fresh forum topic, and
-/// persist the new mapping. Does NOT delete the instance (unlike
-/// [`cleanup_deleted_topic`] which tears down the entire instance).
-///
-/// Returns `Some(new_tid)` on success so callers can retry the send with
-/// the fresh topic. Returns `None` when topic creation fails (no bot
-/// token, network error, etc.) — callers should log and give up.
-pub(crate) fn invalidate_and_recreate_topic(
-    home: &Path,
-    instance_name: &str,
-    stale_tid: i32,
-) -> Option<i32> {
-    tracing::info!(
-        instance = %instance_name,
-        stale_topic_id = stale_tid,
-        "invalidating stale topic and recreating"
-    );
-    unregister_topic(home, stale_tid);
-    // Clear the stale topic_id from fleet.yaml so create_topic_for_instance
-    // doesn't short-circuit on the old value.
-    let _ = crate::fleet::update_instance_field(
-        home,
-        instance_name,
-        "topic_id",
-        serde_yaml_ng::Value::Null,
-    );
-    create_topic_for_instance(home, instance_name)
-}
-
-/// Read the current `agent_state` of `instance_name` from the in-process
-/// [`AgentRegistry`] and return true when the state expects raw keyboard
-/// input rather than inbox-wrapped prose — i.e. `awaiting_operator` (startup
-/// stall) or `interactive_prompt` (pattern-matched modal like codex's update
-/// menu). Returns false when the registry is not attached (daemon bootstrap
-/// not yet wired), the agent is missing, or any lock is poisoned — callers
-/// then fall through to the inbox path rather than dropping messages.
-fn agent_wants_raw_keystrokes(registry: Option<&AgentRegistry>, instance_name: &str) -> bool {
-    let Some(registry) = registry else {
-        return false;
-    };
-    let reg = crate::agent::lock_registry(registry);
-    let Some(handle) = reg.get(instance_name) else {
-        return false;
-    };
-    let core = Arc::clone(&handle.core);
-    // Drop the registry lock before grabbing the per-agent core lock; holding
-    // both at once risks deadlocks against code paths that take core → registry.
-    drop(reg);
-    let guard = core.lock();
-    guard.state.current.wants_raw_keystrokes()
-}
 
 /// Send a reply from an agent to Telegram (called from MCP reply tool).
 #[allow(dead_code)]
@@ -1581,64 +611,6 @@ pub(crate) fn try_telegram_edit(
         .await?;
         Ok::<(), anyhow::Error>(())
     })
-}
-
-/// Reverse-lookup a topic_id for an instance from `topics.json`.
-/// Returns `None` if no mapping exists or the file is missing.
-pub fn lookup_topic_for_instance(home: &std::path::Path, instance_name: &str) -> Option<i32> {
-    let reg = load_topic_registry(home);
-    reg.into_iter()
-        .find(|(_, name)| name == instance_name)
-        .map(|(tid, _)| tid)
-}
-
-/// Create a forum topic for a new instance.
-pub fn create_topic_for_instance(home: &std::path::Path, instance_name: &str) -> Option<i32> {
-    // Idempotent: reuse existing topic from topics.json if present.
-    if let Some(tid) = lookup_topic_for_instance(home, instance_name) {
-        tracing::info!(instance = %instance_name, topic_id = tid, "reusing existing topic");
-        return Some(tid);
-    }
-    let ch = resolve_channel_only_from(home).ok()?;
-    match telegram_runtime().block_on(async {
-        let bot = teloxide::Bot::new(&ch.token);
-        let topic = bot
-            .create_forum_topic(
-                teloxide::types::ChatId(ch.group_id),
-                instance_name,
-                teloxide::types::Rgb::from_u32(0x6FB9F0),
-                "",
-            )
-            .await?;
-        Ok::<i32, anyhow::Error>(topic.thread_id.0 .0)
-    }) {
-        Ok(tid) => {
-            tracing::info!(instance = %instance_name, topic_id = tid, "created topic");
-            register_topic(home, tid, instance_name);
-            Some(tid)
-        }
-        Err(e) => {
-            tracing::error!(instance = %instance_name, error = %e, "failed to create topic");
-            None
-        }
-    }
-}
-
-/// Delete a forum topic.
-pub fn delete_topic(home: &std::path::Path, topic_id: i32) {
-    let ch = match resolve_channel_only_from(home) {
-        Ok(c) => c,
-        Err(_) => return,
-    };
-    let tid = teloxide::types::ThreadId(teloxide::types::MessageId(topic_id));
-    let _ = telegram_runtime().block_on(async {
-        let bot = teloxide::Bot::new(&ch.token);
-        let chat_id = teloxide::types::ChatId(ch.group_id);
-        let _ = bot.close_forum_topic(chat_id, tid).await;
-        bot.delete_forum_topic(chat_id, tid).await
-    });
-    unregister_topic(home, topic_id);
-    tracing::info!(topic_id, "deleted topic");
 }
 
 /// Download an attachment by file_id. `home` scopes credential +
@@ -4002,7 +2974,7 @@ instances:
     /// panics. This grep-based test catches regressions.
     #[test]
     fn handle_message_body_has_no_block_on() {
-        let src = include_str!("telegram.rs");
+        let src = include_str!("inbound.rs");
         // Extract handle_message body (from "async fn handle_message" to next top-level fn)
         let start = src
             .find("async fn handle_message(")
